% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std.R
\name{std}
\alias{std}
\title{Standardize a numeric vector}
\usage{
std(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of equal size to \code{x} containing the standardized
values of the elements of \code{x}.
}
\description{
Standardized the elements of a numeric vector by subtracting off the
arithmetic mean of the vector and dividing by the sample standard deviation
of the vector.
}
\details{
The standard deviation is calculated using the \code{\link[stats:sd]{stats::sd()}}.
}
\examples{
x <- sample(x = 1:100, size = 100, replace = TRUE, prob = rep(0.01, 100))
ssts::std(x)

}
